requirements.txt
streamlit
ultralytics
opencv-python
Pillow
streamlit-webrtc